turno post front

+++ src/main/resources/static/js/delete_odontologo.js
@@ -1,16 +1,17 @@
-function deleteBy(id)
-{
-          //con fetch invocamos a la API de peliculas con el método DELETE
-          //pasandole el id en la URL
-          const url = '/odontologos/eliminar?id='+ id;
-          const settings = {
-              method: 'DELETE'
-          }
-          fetch(url,settings)
-          .then(response => response.json())
+function deleteBy(id){
 
-          //borrar la fila de la pelicula eliminada
-          let row_id = "#tr_" + id;
-          document.querySelector(row_id).remove();
+    if(window.confirm("¿Está seguro de eliminar el odontologo con id: " + id +"?")){
+    //con fetch invocamos a la API de peliculas con el método DELETE
+    //pasandole el id en la URL
+    const url = 'odontologos/eliminar/'+ id;
+    const settings = {
+        method: 'DELETE'
+    }
+    fetch(url,settings)
+    .then(response => response.json())
 
+    //borrar la fila de la pelicula eliminada
+    let row_id = "#tr_" + id;
+    document.querySelector(row_id).remove();
+  }
 }
